Fix renaming to quoted names on SQL Server
sp_rename accepts the old name as a qualified name but the new name as
a literal value.

The caller of the stored procedure should be able to qualify the old
name in order to reference a column in some table or a table in some
schema, etc. But it's not necessary for the new name because the
reference to the object being renamed is already established in the
old name.