bugfix: deallocate mysqli prepared statement (#6681)
Long running processes might hit the `max_prepared_stmt_count` due not
deallocating the statement correctly.

<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | bug
| Fixed issues | -

#### Summary

For postgres there was an implementation to deallocate prepared
statements to prevent hitting a any limits. This was not the case for
the `mysqli` drivers. Which can suffer from the same problem.

In this PR a `mysqli_stmt_close` is added to properly cleanup any open
prepared statement.

A similar approach was implemented for PostgreSQL in this PR
https://github.com/doctrine/dbal/pull/5893

---------

Co-authored-by: Alexander M. Turek <me@derrabus.de>