Remove @throws Exception from _getPortable*Definition() signatures
Each of these methods effectively parses the results of the
corresponding schema introspection, so the API consumer cannot do
anything to handle an exception that occurred during parsing (e.g.
unsupported column data type).

If such a method performs other queries to the database which may
result in a run-time exception, it violates the contract and needs to be
reworked.