Convert driver exceptions when starting transactions
For long-running processes, it's possible that the wrapped connection is
no longer valid and starting a transaction will lead to driver
exceptions.

DBAL Connection wasn't converting these, leaking internal details of the
platform and not providing a consistent API (all other methods would
throw a `ConnectionLost` exception for the scenario above).

This ensures we do have the expected behaviour for this edge-case.

Signed-off-by: Luís Cobucci <lcobucci@gmail.com>