Only trim the leading slash from the DSN if a host is also available (#6790)
|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | #6789

#### Summary

Previously the DsnParser unconditionally removed the first character
from the path, without verifying that it is actually a slash and that it
is dealing with a reasonably well-formed URI. This causes confusing
behavior when a malformed URI that survives `parse_url()` is passed to
`parse()`, since DsnParser would return the input as the `dbname`, but
with the first character missing.

This can lead to hard-to-debug situations when the DSN is coming from an
environment variable, where quoting characters are accidentally included
in the value itself, instead of being interpreted by the tool setting
the env, since the `dbname` almost looks like the intended input (but
still having the trailing quote).

see doctrine/dbal#6789