fix(PgSQL): Allow to pass IPv6 address in URI notation for postgres (#6344)
<!-- Fill in the relevant information below to help triage your pull
request. -->

|      Q       |   A
|------------- | -----------
| Type         | improvement
| Fixed issues | <!-- use #NUM format to reference an issue -->

#### Summary

Allow to pass IPv6 address of postgres server in URI notation
(`[ff:aa:...]`).
When the host is passed in URI format to `host` parameter of
`pg_connect` it will fail because it then tries to resolve it as when it
was a real host name.
Instead just pass the IP address to the `hostaddr` parameter.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>
Co-authored-by: Alexander M. Turek <me@derrabus.de>